[Ashwini] enabled apple sign in
