+++ WriteUp/Helpers/Extension.swift
@@ -58,12 +58,12 @@ extension UIFont {
     
     func itemTitle(size: CGFloat? = 17) -> UIFont {
         return UIFont(name: "MaisonNeue-Demi", size: 17)!
-        //UIFont.systemFont(ofSize: 17, weight: .medium)
+//        UIFont.systemFont(ofSize: 17, weight: .medium)
     }
     
     func itemDesc(size: CGFloat? = 15) -> UIFont {
         return UIFont(name: "MaisonNeue-book", size: 14)!
-        //UIFont.systemFont(ofSize: 15, weight: .regular)
+//        UIFont.systemFont(ofSize: 15, weight: .regular)
     }
     
     func formControlTitle(size: CGFloat? = 17) -> UIFont{
@@ -72,7 +72,7 @@ extension UIFont {
     
     func formControlSegmented(size: CGFloat? = 13) -> UIFont {
         return UIFont(name: "MaisonNeue-Bold", size: size!)!
-        //return UIFont.systemFont(ofSize: 13, weight: .regular)
+//        return UIFont.systemFont(ofSize: 13, weight: .regular)
     }
     
     func textInput(size: CGFloat? = 17) -> UIFont {
@@ -81,7 +81,7 @@ extension UIFont {
     
     func tabBarTitle(size: CGFloat? = 10) -> UIFont {
         return UIFont(name: "MaisonNeue-Light", size: size ?? 10)!
-        //return UIFont.systemFont(ofSize: 10, weight: .medium)
+//        return UIFont.systemFont(ofSize: 10, weight: .medium)
     }
 }
 
@@ -208,39 +208,39 @@ extension String {
 
 }
 
-#warning("Move Image extension to separate file")
-extension UIImage {
-    
-    struct IconConfig {
-        static let navbar = UIImage.SymbolConfiguration(weight: UIImage.SymbolWeight.bold)
-    }
-    
-    struct ImageConfig {
-        #warning("Need to create config for bigger image")
-        static let profilePlaceHolder = UIImage.SymbolConfiguration(weight: UIImage.SymbolWeight.bold)
-    }
-    
-    struct Common {
-        static let empty = UIImage()
-        static let newNote = UIImage(named: Constant.HeaderNoteBar.addNoteImage)
-        static let profilePlaceHolder = UIImage(systemName: "person.crop.circle", withConfiguration: UIImage.ImageConfig.profilePlaceHolder)
-    }
-    
-    struct NavBarIcon {
-        static let createNote = UIImage(systemName: "doc.badge.plus", withConfiguration: UIImage.IconConfig.navbar)
-        static let profile = UIImage(systemName: "person.crop.circle", withConfiguration: UIImage.IconConfig.navbar)
-    }
-    
-    struct ToolBarIcon {        
-        static let deleteNote = UIImage(systemName: "trash.circle", withConfiguration: UIImage.IconConfig.navbar)
-        static let clearText = UIImage(systemName: "xmark.circle", withConfiguration: UIImage.IconConfig.navbar)
-    }
-    
-    struct Image {
-        static let smallSymbolImage = UIImage(systemName: "person.crop.circle", withConfiguration: UIImage.ImageConfig.profilePlaceHolder)
-    }
-    
-    struct Main {
-        static let largeAppIcon = UIImage(named: Constant.SignInSC.logoImageName)
-    }
-}
+//#warning("Move Image extension to separate file")
+//extension UIImage {
+//    
+//    struct IconConfig {
+//        static let navbar = UIImage.SymbolConfiguration(weight: UIImage.SymbolWeight.bold)
+//    }
+//    
+//    struct ImageConfig {
+//        #warning("Need to create config for bigger image")
+//        static let profilePlaceHolder = UIImage.SymbolConfiguration(weight: UIImage.SymbolWeight.bold)
+//    }
+//    
+//    struct Common {
+//        static let empty = UIImage()
+//        static let newNote = UIImage(named: Constant.HeaderNoteBar.addNoteImage)
+//        static let profilePlaceHolder = UIImage(systemName: "person.crop.circle", withConfiguration: UIImage.ImageConfig.profilePlaceHolder)
+//    }
+//    
+//    struct NavBarIcon {
+//        static let createNote = UIImage(systemName: "doc.badge.plus", withConfiguration: UIImage.IconConfig.navbar)
+//        static let profile = UIImage(systemName: "person.crop.circle", withConfiguration: UIImage.IconConfig.navbar)
+//    }
+//    
+//    struct ToolBarIcon {        
+//        static let deleteNote = UIImage(systemName: "trash.circle", withConfiguration: UIImage.IconConfig.navbar)
+//        static let clearText = UIImage(systemName: "xmark.circle", withConfiguration: UIImage.IconConfig.navbar)
+//    }
+//    
+//    struct Image {
+//        static let smallSymbolImage = UIImage(systemName: "person.crop.circle", withConfiguration: UIImage.ImageConfig.profilePlaceHolder)
+//    }
+//    
+//    struct Main {
+//        static let largeAppIcon = UIImage(named: Constant.SignInSC.logoImageName)
+//    }
+//}
